fix(auth): improve error handling for login and registration
